# packages/backend/api-spec.yaml
openapi: "3.0.3"
info:
  title: ATB AgriTrace API
  version: "1.0.0"
  description: API de traçabilité agricole blockchain
servers:
  - url: https://api.agritrace.bj
    description: Production
  - url: http://localhost:8000
    description: Local dev
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /api/lots/{id}:
    get:
      tags: [Lots]
      summary: Détails d'un lot
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: Détails du lot }
...